refactor: move lecturer related api call to its own file
